pyproject.toml
[tool.poetry]
name = "mything"
version = "0.2.2"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = "19.10b0"
transcrypt = "^3.7"
pytest-doctestplus = "^0.5.0"
pylint = "^2.4"
dependencies = "^0.15.0"
pytest-watch = "^4.2"
pytest-bdd = "^3.2"
pytest-splinter = "^2.0.1"
pytest-benchmark = "^3.2.2"
flask = "^1.1.1"
lxml = "^4.4.2"
cssselect = "^1.1.0"

[tool.black]
line-length = 110
skip-string-normalization = true
exclude = "docs"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
